+++ code/core/src/main/java/org/betonquest/betonquest/config/patcher/migration/QuestMigration.java
@@ -1,5 +1,6 @@
 package org.betonquest.betonquest.config.patcher.migration;
 
+import org.betonquest.betonquest.api.bukkit.config.custom.multi.MultiConfiguration;
 import org.betonquest.betonquest.config.quest.Quest;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.InvalidConfigurationException;
@@ -71,6 +72,82 @@ public interface QuestMigration {
             if (value != null && filter.apply(value)) {
                 section.set(key, operation.apply(value));
             }
+        }
+    }
+
+    /**
+     * Renames a configuration section in the config.
+     *
+     * @param config  the source config
+     * @param oldPath the section path to rename
+     * @param newPath the new name of the section
+     * @throws InvalidConfigurationException if an error occurs
+     */
+    default void renameSection(final MultiConfiguration config, final String oldPath, final String newPath) throws InvalidConfigurationException {
+        final ConfigurationSection oldSection = config.getConfigurationSection(oldPath);
+        if (oldSection == null) {
+            return;
+        }
+        final ConfigurationSection newSection = config.createSection(newPath);
+        for (final String key : oldSection.getKeys(false)) {
+            final ConfigurationSection source = config.getSourceConfigurationSection(oldPath + "." + key);
+            if (source == null) {
+                throw new InvalidConfigurationException("Cannot migrate '" + oldPath + "' to '" + newPath + "' for key: " + key);
+            }
+            newSection.set(key, oldSection.get(key));
+            newSection.setComments(key, oldSection.getComments(key));
+            newSection.setInlineComments(key, oldSection.getInlineComments(key));
+            config.associateWith(newPath + "." + key, source);
+        }
+        config.set(oldPath, null);
+    }
+
+    /**
+     * Renames a key in all subsections of the given path configuration key in a specific section.
+     *
+     * @param root   the source configuration of the path and associate with
+     * @param path   the base path of the sections to replace the keys in
+     * @param oldKey the old to replace key
+     * @param newKey the new key
+     * @throws InvalidConfigurationException if an error occurs
+     * @see #replaceKeyInSection(MultiConfiguration, ConfigurationSection, String, String)
+     */
+    default void replaceKeyInSections(final MultiConfiguration root, final String path, final String oldKey, final String newKey)
+            throws InvalidConfigurationException {
+        final ConfigurationSection config = root.getConfigurationSection(path);
+        if (config == null) {
+            return;
+        }
+        for (final String key : config.getKeys(false)) {
+            final ConfigurationSection section = config.getConfigurationSection(key);
+            if (section == null) {
+                continue;
+            }
+            replaceKeyInSection(root, section, oldKey, newKey);
+        }
+    }
+
+    /**
+     * Renames a configuration key in a specific section.
+     *
+     * @param root    the source configuration to associate with
+     * @param section the section to replace the key in
+     * @param oldKey  the old to replace key
+     * @param newKey  the new key
+     * @throws InvalidConfigurationException if an error occurs
+     */
+    default void replaceKeyInSection(final MultiConfiguration root, final ConfigurationSection section,
+                                     final String oldKey, final String newKey) throws InvalidConfigurationException {
+        final String value = section.getString(oldKey);
+        if (value != null) {
+            final ConfigurationSection source = root.getSourceConfigurationSection(section.getCurrentPath() + "." + oldKey);
+            if (source == null) {
+                throw new InvalidConfigurationException("Cannot migrate key '" + oldKey + "' to '" + newKey + "'");
+            }
+            section.set(newKey, value);
+            section.setComments(newKey, section.getComments(oldKey));
+            section.setInlineComments(newKey, section.getInlineComments(oldKey));
+            section.set(oldKey, null);
         }
     }
 }