+++ code/core/src/main/java/org/betonquest/betonquest/quest/event/notify/NotifyEventFactory.java
@@ -133,7 +133,7 @@ public class NotifyEventFactory implements PlayerEventFactory {
             data.put(key, value);
         }
 
-        data.remove("events");
+        data.remove("actions");
         data.remove("conditions");
 
         return data;